Changed the way I get environmental variables and fixed some issues
Re-factored the code, so it consumes the MongoDB connection string in a more secure way.
Fixed a issue with the retreived objects from MongoDB having members set to null. identified a small issue with {get; set;} properties on the RecipeTitle and Recipe Description members of the Recipe class. Fixed the issue, by adding a property init, as set or init was missing.
Modified the insert and delete methods of the MongoDBClient class, as they were taking in an argument "collection" which in this case was unnecessary.