Implemented very basic authentication with JWT
Implemented authentication with very basic useage of JWT.
Added an AuthController which is responsible for handling the authentication. The controller has a basic method called SignIn, that takes as an argument a test class object called AuthDTO. Just for simplicity, at the current point no information is encrypted, the password is saved in plain text. The controller also holds the method for generating JWT, which is to be moved to a separate controller later.
Currently the SignIn method just checks if the password equals a hardcoded value.