Added UserService service for actual authentication
Changed the structure and logic of how authentication works, instead of checking for hardcoded values, the Auth controller is utilizing the UserService to look up the user credentials in the MongoDB database.
The UserService class currently has two methods, one for SignIn and one for SignUp. The SignIn method looks up the database for matches of documents with the exact same cred. as the ones being passed and if so, returns the object, if no, returns null and thus lets the controller know, there is no user with these cred. The SignUp method similarly to the SignIn takes in to arguments , email and password., but returns a Boolean The method first looks up the database if there are other profiles with this email registered, if they are, returns false, if there are non, calls the InsertOneAsync method part of the MongoDB client and returns true, to let the controller know the actions was successful.