Added RecipeService and a little bit of Mark up in the MVC app and strated building the update routes in the API
Added RecipeService which will be in charge of handling all interactions with recipes to and from the API.  Altered the index view of the application as well, to display all recipes in the database in a table. The Home controller feeds the Index view the fetched data and the Index view displays them on screen in table form
Added UpdateRecipe controller in the API, that is to be finished. Also wrote an IUserService interface to follow best practices. Added a new method in the MongoDB client. All of the methods from the Mongo DB client are to be moved to a separate service.